+++ xoeuf/tests/test_enumeration/tests/__init__.py
@@ -3,24 +3,17 @@
 # ---------------------------------------------------------------------
 # __init__
 # ---------------------------------------------------------------------
-# Copyright (c) 2016-2017 Merchise Autrement [~º/~] and Contributors
+# Copyright (c) 2017 Merchise Autrement [~º/~] and Contributors
 # All rights reserved.
 #
 # This is free software; you can redistribute it and/or modify it under the
 # terms of the LICENCE attached (see LICENCE file) in the distribution
 # package.
 #
-# Created on 2016-11-15
 
 from __future__ import (division as _py3_division,
                         print_function as _py3_print,
                         absolute_import as _py3_abs_import)
 
 
-from xoeuf.odoo.fields import *          # noqa: reexport
-
-from .datetime import LocalizedDatetime  # noqa: reexport
-from .properties import Property         # noqa: reexport
-from .monetary import Monetary           # noqa: reexport
-from .timespan import TimeSpan           # noqa: reexport
-from .enumeration import Enumeration     # noqa: reexport
+from . import test_enumeration